+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Add missing block-only signature for debug logging
 - update (bundled) rack to 2.2.20
 - Ensure rack boot process leaves ENV['GEM_PATH'] and Gem.paths in a consistent state
+- Remove undocumented and unsafe jruby.rack.env.gem_path = false option (unusable on Bundler 1.6+)
 
 ## 1.2.5